feat: Add Gradio frontend, MNIST/LeNet support, and CIM tensor cast fix
inference.py
- New --run_dir to flatten artifact directory (skips deep make_path tree)
- Save accuracy.json next to logs
- Tee C++ PPA stdout into ppa_output.log instead of dropping to stdout
- Dispatch MNIST via the new get_mnist loader

dataset.py
- get_mnist() loader with standard 1x28x28 normalization

quantize.py
- Wire LeNet (mnist) into the model-selection branch

models/lenet.py
- LeNet-5 (Conv-ReLU-Pool x2 -> FC x3) using only ops covered by the
  CIM quant map so no extra registration is needed

models/vgg.py
- Load checkpoints with map_location='cpu' for CUDA/PyTorch portability

pytorch-quantization/.../macro.py
- Cast weight index tensors to .long() before advanced indexing into
  mem_states (PyTorch requires long/bool/byte for index tensors)

gradio_app.py
- Web UI wrapping inference.py: per-arg form, real-time log streaming,
  CSV uploads for mem_states / output_noise, log + results.zip download,
  Stop button, atexit + PDEATHSIG cleanup so the GPU child dies with the
  parent on any termination

train_mnist_lenet.py
- Tiny SGD/MultiStepLR trainer that writes models/lenet_mnist.pth in
  the format quantize.py expects

mem_states_4bit*.csv
- 16-state (bitcell=4) example device profiles (ideal + 30% sigma)

NeuroSIM/NetWork_lenet.csv
- Topology snapshot regenerated by hook.make_records during a LeNet run

+++ pytorch-quantization/pytorch_quantization/cim/modules/macro.py
@@ -196,8 +196,11 @@ class CIM():
             return torch.clamp(torch.round(ADC_out), min=0.0, max= float(2**self._cim_args.adc_precision))
         
         # Get memory states
-        weights = weights.to(self._cim_args.mem_states.device)
-        weights_ref = weights_ref.to(self._cim_args.mem_states.device)
+        # NOTE: PyTorch advanced indexing requires index tensors to be
+        # torch.long / bool / byte. The compute path in simulate_array uses
+        # int32 for bit slicing, so cast here for safe indexing.
+        weights = weights.to(self._cim_args.mem_states.device).long()
+        weights_ref = weights_ref.to(self._cim_args.mem_states.device).long()
         
         memory_states = self._cim_args.mem_states[weights]
         memory_states_ref = self._cim_args.mem_states[weights_ref]
